Move _visit_child_role to DelegatedRole
Make _visit_child_role a public method of
DelegatedRole class. Reduce debug logging.
